OnInterest: registeredPrefixId should be long.

+++ src/net/named_data/jndn/OnInterest.java
@@ -26,5 +26,5 @@ public interface OnInterest {
    */
   void onInterest
     (Name prefix, Interest interest, Transport transport, 
-     int registeredPrefixId);
+     long registeredPrefixId);
 }
